extract TypeArgumentVisitor to a separat class
